\name{reverseGeo-package}
\alias{reverseGeo-package}
\alias{reverseGeo}
\docType{package}
\title{
\packageTitle{reverseGeo}
}
\description{
\packageDescription{reverseGeo}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{reverseGeo}
\packageIndices{reverseGeo}

}
\author{
\packageAuthor{reverseGeo}

Maintainer: \packageMaintainer{reverseGeo}
}


\keyword{ package }
\examples{
~~ simple examples of the most important functions ~~
}
