\name{reverseGeo}
\alias{reverseGeo}
\title{Reverse geocoding utility}
\description{An interface for the reverse geocoding service by Google Maps Geocoding API's. 
	Reverse geocoding is the process of converting geographic coordinates into a human-readable address.}
\usage{
reverseGeo(lat, lon, key)
}

\arguments{
  \item{lat}{Latitude}
  \item{lon}{Longitude}
  \item{key}{Your application's API Google Maps key}
}

\value{A human-readable address}

\author{Alejandro Platas}

