\name{reverseGeo-package}
\alias{reverseGeo-package}
\alias{reverseGeo}
\docType{package}
\title{
\packageTitle{reverseGeo}
}
\description{
\packageDescription{reverseGeo}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{reverseGeo}
\packageIndices{reverseGeo}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{reverseGeo}

Maintainer: \packageMaintainer{reverseGeo}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
