\name{reverseGeo-package}

\alias{reverseGeo-package}

\alias{reverseGeo}

\docType{package}

\title{An interface for the reverse geocoding service by Google Maps Geocoding API's.}

\description{An interface for the reverse geocoding service by Google Maps Geocoding API's. 
	Reverse geocoding is the process of converting geographic coordinates into a human-readable address.}

\details{

\tabular{ll}{Package: \tab reverseGeo\cr

Type: \tab Package\cr

Title: \tab An interface for the reverse geocoding service by Google Maps Geocoding API's.\cr

Version: \tab 1.1.0\cr

Date: \tab 2017-25-07\cr

Depends: \tab R (>= 2.10)\cr

Author: \tab Alejandro Platas-López\cr

Maintainer: \tab Alejandro Platas-López <alejandroplatasl@gmail.com>\cr

License: \tab GPL\cr

LazyLoad: \tab yes\cr}

}

\author{Alejandro Platas-López}



\keyword{ package }
